Layout and Functionality
Added the Layout fixed so that the timer and the button are not placed ontop of each other. Additionally made it so that the timer stops after the person had correctly solved the puzzle. With the functionality I added the ability to choose difficulty through the terminal running the program. This will be changed to be incorporated through the menu.py file when I create the main menu for the project.