UI Cleaning and Login Page
UI is cleaned up so I can edit the positions I want the current UI I have to be in. Additionally, I added the login feature UI, with the functionality of being able to enter the details. Currently the login button does not work. Also created the difficulty_selection.py file which will house the UI for the user to select the difficulty of the sudoku that they want. Deleted test_file.py as it no longer had any use for testing purposes. Removed that from the .gitignore.