Account Creation
Made it so the user is able to create a new account. Currently this does nothing as the login feature is not finished. But it still allows the user to have an email username and password saved to the data base. Additionally, started working on the difficulty selection with planning the design of it.