Sudoku Manager Class
Started working on the implementation of the Sudoku Manager class, which will allow me to seperate the UI from the actual board which will make it easier to change values and implement new features. This is still being worked on but currently works for the board creation and the hint button.