Difficulty Selection
Added the feature of being able to see your best times next to each difficulty to difficulty_selection.py file. This only works if the user has logged in. If user is not logged in, they see a black opace box covering the times informing them to sign in to get access to the feature.