Game End Update
Added the feature that adds a popup that appears when the user has successfully completed the sudoku board. Next if the time is a best time, which means the amount of time taken to complete the puzzle is less than their previous best for the difficulty, or if their time is 0, then the database will update their scores accordingly. This will also show on the popup telling them that they have achieved a new best time. The popup contains two buttons, one of which says Play Again, which takes the user to the difficulty selection page which will show them their best times with each difficulty. And the other button takes the user back to the main menu. I also ensured to make the hint limit 3 hints and made it reset whenever the user went into another puzzle. I also added the convert_time_to_seconds function that takes a string i the format of "mm:ss" and converts it into an integer in seconds. This is the value that it uses to compare best times.