Minor Improvements
Added the error_message to the create account menu which will now show if an invalid email is entered. Additionally, added the returning of True or False if there is an invalid email, this is the value used to determine what error message needs to be shown in the account creation menu.