Completed Sudoku interaction
Added a popup when the user completes the sudoku, which tells them their time, their personal best time, the time between their current time and personal best time, and two buttons which allows the user to go to the main menu or the difficulty selection screen. Made it so that the users times get updates to the database whenever they achieve a new best score. Added an unsolved message that appears if the user clicks check solve button and puzzle is unsolved. Added the function that converts a time given in seconds to the format "00:00" e.g. turns 105 into "01:45" which allows me to compare the times and easily find out how much time between the users scores there were.