Password Changing
Added the functionality that allows the user to change their password from the account menu. Had to add functions to the databasemanager.py file which checks if the password they entered in their original password position is the password that they currently have. Additionally, It checks to ensure that the user has their new password confirmed.